refactor: rename DOMAIN_CONFIG_TYPE to DOMAIN_CONFIG_NAMES across files
