+++ apps/web/src/common/modules/navigations/top-bar/constants/constant.ts
@@ -1,11 +1,11 @@
-import { DOMAIN_CONFIG_TYPE } from '@/store/domain/constant';
+import { DOMAIN_CONFIG_NAMES } from '@/store/domain/constant';
 
 import type { MenuId } from '@/lib/menu/config';
 import { MENU_ID } from '@/lib/menu/config';
 
 export const customMenuNameList: MenuId[] = [
     MENU_ID.DASHBOARDS,
-    DOMAIN_CONFIG_TYPE.EXTRA_MENU as MenuId,
+    DOMAIN_CONFIG_NAMES.EXTRA_MENU as MenuId,
 ];
 
 export const WORKSPACE_LOGO_ICON_THEMES = ['blue', 'yellow', 'gray', 'green', 'coral', 'indigo', 'peacock'] as const;
