+++ apps/web/src/store/domain/type.ts
@@ -1,4 +1,4 @@
-import type { DOMAIN_CONFIG_TYPE } from '@/store/domain/constant';
+import type { DOMAIN_CONFIG_NAMES } from '@/store/domain/constant';
 
 export type ExtendedAuthType = 'GOOGLE_OAUTH2' | 'KEYCLOAK_OIDC' | 'SAML';
 
@@ -18,4 +18,4 @@ export interface ExtraSubMenu {
     label: string;
 }
 
-export type DomainConfigKey = keyof typeof DOMAIN_CONFIG_TYPE;
+export type DomainConfigKey = keyof typeof DOMAIN_CONFIG_NAMES;
